+++ Sources/GanitEditorUI/SheetCommands.swift
@@ -17,6 +17,7 @@ import AppKit
   func nextProblem(_ sender: Any?)
   func previousProblem(_ sender: Any?)
   func insertSubtotal(_ sender: Any?)
+  func insertAnswerArrow(_ sender: Any?)
   func toggleHeading(_ sender: Any?)
   func toggleComment(_ sender: Any?)
   func insertDivider(_ sender: Any?)